Refactor CLI helper functions into utils module
